+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## 7.0.0 (2022-??-??)
+## 7.0.0 (2022-12-25)
 
-* Raised minimum required PHP version to 8.0.0
+* Raised minimum required PHP version to 8.0
 * Added support for intersection types
 * Added support for combined union and intersection types